Move members test to correct file.
